requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==0.6.8
# dbt-core
# dbt-postgres
# dbt-snowflake
# airflow-dbt
apache-airflow-providers-snowflake==2.6.0
apache-airflow-providers-apache-spark==3.0.0